requirements.txt
fastapi==0.115.12
uvicorn==0.34.2
gunicorn
azure-cosmos==4.9.0
pydantic==2.11.4
dotenv==0.9.9
